requirements.txt
tensorflow==2.18.0
keras==3.8.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.2.2
matplotlib==3.7.2
pillow==11.2.1
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.9
gdown==5.2.0
tensorflow==2.18.0
keras==3.8.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.2.2
matplotlib==3.7.2
pillow==11.2.1
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.9
gdown==5.2.0
streamlit